Refactor file structure and add BookGalleryPage
