+++ frontend/src/App.tsx
@@ -3,9 +3,9 @@ import "./types/types.ts"
 import axios from "axios"
 import {Book} from "./types/types.ts";
 import {useEffect, useState} from "react";
-import BookGallery from "./components/bookGallery/BookGallery.tsx";
-import {Route, Routes} from "react-router-dom";
-import BookDetailsPage from "./components/bookDetailsPage/BookDetailsPage.tsx";
+import {Link, Route, Routes} from "react-router-dom";
+import BookDetailsPage from "./pages/BookDetailsPage/bookDetailsPage/BookDetailsPage.tsx";
+import BookGalleryPage from "./pages/BookGalleryPage/bookGalleryPage/BookGalleryPage.tsx";
 
 
 function App() {
@@ -32,8 +32,9 @@ function App() {
     <>
         <h1>LibraryApp</h1>
         <h3>Cool motto ( ͡° ͜ʖ ͡°)</h3>
+        <Link to={"/books"}>All Books</Link>
         <Routes>
-            <Route path={"/"} element={<BookGallery data={data}/>}/>
+            <Route path={"/books"} element={<BookGalleryPage data={data}/>}/>
             <Route path={"/books/:id"} element={<BookDetailsPage/>}/>
         </Routes>
